Move `channel_type::Error` to `error::ChannelTypeError`
