+++ src/error.rs
@@ -1,2 +1,28 @@
 use thiserror::Error;
 
+#[derive(Error, Eq, PartialEq, Clone, Debug)]
+pub enum ChannelTypeError {
+    // Marshal buffer was too short
+    UnexpectedEndOfBuffer { expected: usize, actual: usize },
+
+    // Remote requested a channel type that we don't support
+    InvalidChannelType { invalid_type: u8 },
+}
+
+impl std::fmt::Display for ChannelTypeError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::UnexpectedEndOfBuffer { expected, actual } => {
+                writeln!(
+                    f,
+                    "Marshal buffer was too short: (expected: {:?}, actual: {:?})",
+                    expected, actual
+                )
+            }
+            Self::InvalidChannelType { invalid_type } => {
+                writeln!(f, "Invalid channel type: {:?}", invalid_type)
+            }
+        }
+    }
+}
+
